+++ packages/common/web/src/Base/createBaseComponent.jsx
@@ -2,7 +2,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { findDOMNode } from 'react-dom';
-import { isNil } from 'lodash';
+import { compact, isNil } from 'lodash';
 import classNames from 'classnames';
 import createHocName from '../Utils/createHocName';
 import getDisplayName from '../Utils/getDisplayName';
@@ -245,11 +245,12 @@ const createBaseComponent = ({
       const { isFocused } = blockProps;
       const isActive = isFocused && isEditorFocused && !readOnly;
 
-      const classNameStrategies = [
+      const classNameStrategies = compact([
         PluginComponent.alignmentClassName || alignmentClassName,
         PluginComponent.sizeClassName || sizeClassName,
         PluginComponent.textWrapClassName || textWrapClassName,
-      ].map(strategy => strategy(this.state.componentData, theme, this.styles, isMobile));
+        PluginComponent.customClassName,
+      ]).map(strategy => strategy(this.state.componentData, theme, this.styles, isMobile));
 
       const ContainerClassNames = classNames(
         {