+++ packages/plugin-divider/web/src/typeMapper.js
@@ -1,6 +1,14 @@
 import DividerComponent from './components/divider-component';
 import { DIVIDER_TYPE } from './constants';
+import { customClassName } from './classNameStrategies';
+
+const dividerRenderDescriptor = {
+  component: DividerComponent,
+  classNameStrategies: {
+    custom: customClassName,
+  },
+};
 
 export const typeMapper = () => ({
-  [DIVIDER_TYPE]: { component: DividerComponent },
+  [DIVIDER_TYPE]: dividerRenderDescriptor,
 });