+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - `image`
   - [75336e34](https://github.com/wix-incubator/rich-content/commit/75336e34) open imageStudio on crop tab
   - [#413](https://github.com/wix-incubator/rich-content/pull/413) new image doesn't replace existing image
+  - [#428](https://github.com/wix-incubator/rich-content/pull/428) image settings works while uploading
   - [#423](https://github.com/wix-incubator/rich-content/pull/423) images are rendered in high resolution for seo
 - `file-upload`
   - [#426](https://github.com/wix-incubator/rich-content/pull/426) apply spinner animation to file upload plugin