+++ CHANGELOG.md
@@ -14,10 +14,13 @@
     Click to see more.
   </summary>
 
+  ### :rocket: New Feature
+  - `video`
+    - [#443](https://github.com/wix-incubator/rich-content/pull/443) async url resolving support
   ### :bug: Bug Fix
   - `viewer`
     - [#442](https://github.com/wix-incubator/rich-content/pull/442) html plugin alignment
-
+    
 </details>
 <hr/>