+++ CHANGELOG.md
@@ -18,6 +18,8 @@
   - `video`
     - [#443](https://github.com/wix-incubator/rich-content/pull/443) async url resolving support
   ### :bug: Bug Fix
+  - `divider`
+    - [#438](https://github.com/wix-incubator/rich-content/pull/438) container alignment
   - `viewer`
     - [#442](https://github.com/wix-incubator/rich-content/pull/442) html plugin alignment
     
@@ -51,7 +53,7 @@
   - [#434](https://github.com/wix-incubator/rich-content/pull/434) text dropdown buttons tooltips display
   - [#433](https://github.com/wix-incubator/rich-content/pull/433) side toolbar is positioned to the right in RTL
 - `divider`
-  - [#406](https://github.com/wix-incubator/rich-content/pull/406) divider center alignment
+  - [#406](https://github.com/wix-incubator/rich-content/pull/406) center alignment
 - `image`
   - [75336e34](https://github.com/wix-incubator/rich-content/commit/75336e34) open imageStudio on crop tab
   - [#413](https://github.com/wix-incubator/rich-content/pull/413) new image doesn't replace existing image