+++ CHANGELOG.md
@@ -78,7 +78,7 @@
   - [#387](https://github.com/wix-incubator/rich-content/pull/387) data validation schema added
 <hr/>
 
-## 5.1.11 (Nov 4, 2019)
+## 5.1.12 (Nov 4, 2019)
 ### :house: Internal
   - `gallery`
     - [#446](https://github.com/wix-incubator/rich-content/pull/446) `pro-gallery` updated to `v1.5.25`